beginning some refactoring

+++ Assets/Scripts/Combat/Abilities/AttackAbility.cs
@@ -5,7 +5,7 @@ public class AttackAbility : Ability
 {
     public AttackAbility(AbilityData data, GameObject source, List<StatusEffectData> statusEffectDataList) : base(data, source, statusEffectDataList) { }
 
-    public virtual void Trigger(CombatUnit user, CombatUnit target, out int rawDamage)
+    public virtual void Trigger(CombatUnit user, CombatUnit target, out float rawDamage)
     {
         AttackAbilityData data = abilityData as AttackAbilityData;
         // can add TriggerEffect method to add vfx
