+++ Assets/Scripts/Combat/Abilities/AttackAbilityData.cs
@@ -8,11 +8,11 @@ using UnityEngine;
 public class AttackAbilityData : AbilityData
 {
 
-    public delegate int del(CombatUnit user, CombatUnit target);
+    public delegate float del(CombatUnit user, CombatUnit target);
     public del DamageFunc = delegate(CombatUnit user, CombatUnit target)
     {
-        int userAtk = user.GetAttack();
-        int targetDef = target.GetDefense();
+        float userAtk = user.GetAttack();
+        float targetDef = target.GetDefense();
 
         return userAtk * 2 - targetDef;
     };
